fix: otter making repeated api calls
The useSettings hook repeated calls to settings API after a recent version, probably after AI Content creator as that hooks to function to each block.

After trying various methods, I settled on this one as it seems to work with all the current usage of useSettings that we have in Otter.

I also tried to use a better implementation but those seem to not always work with the various use cases that we have with Otter. (as in using the Data Store for saving/retrieving the data)

In the future, I'd avoid using useSettings in plugins as now Gutenebrg has built-in methods to use Settings API from the store, and as we work on different parts of Otter, we can try to remove the existing usage of useSettings.